Library can now send multicast QUIC packets
This is a large commit which makes the faked server handshake logic work.
The fake packets can be found in lib/multicast.h, and are fed into ngtcp2
to make it think it's actually communicating with another peer.

Using the example multicast-sender (was example-server), the library will
now send a PUSH_PROMISE, HEADERS and DATA frame to a multicast group.