RD992 - Add callback-based timer mechanism to fix stalled streams
The old ngtcp2 version of the library used to use these callback-based
timers, but the interfaces weren't removed for this ngtcp2-less version
even though it's not actually used by the library. This then reintroduces
this functionality for the purposes of timing out streams or whole
sessions that have lain idle for too long. There is an exception for
stream ID 0, as this is tied to the whole session timeout.